Liberty Release Candidate 1

Conflicts:
	mistral/yaql_utils.py

+++ mistral/api/controllers/v2/workflow.py
@@ -1,5 +1,6 @@
 # Copyright 2013 - Mirantis, Inc.
 # Copyright 2015 - StackStorm, Inc.
+# Copyright 2015 Huawei Technologies Co., Ltd.
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -13,6 +14,7 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
+from oslo_log import log as logging
 import pecan
 from pecan import hooks
 from pecan import rest
@@ -20,10 +22,11 @@ from wsme import types as wtypes
 import wsmeext.pecan as wsme_pecan
 
 from mistral.api.controllers import resource
+from mistral.api.controllers.v2 import types
 from mistral.api.controllers.v2 import validation
 from mistral.api.hooks import content_type as ct_hook
 from mistral.db.v2 import api as db_api
-from mistral.openstack.common import log as logging
+from mistral import exceptions as exc
 from mistral.services import workflows
 from mistral.utils import rest_utils
 from mistral.workbook import parser as spec_parser
@@ -70,15 +73,16 @@ class Workflow(resource.Resource):
             if hasattr(e, key):
                 setattr(e, key, val)
 
-        input = d['spec'].get('input', [])
-        for param in input:
-            if isinstance(param, dict):
-                for k, v in param.items():
-                    input_list.append("%s=%s" % (k, v))
-            else:
-                input_list.append(param)
+        if 'spec' in d:
+            input = d.get('spec', {}).get('input', [])
+            for param in input:
+                if isinstance(param, dict):
+                    for k, v in param.items():
+                        input_list.append("%s=%s" % (k, v))
+                else:
+                    input_list.append(param)
 
-        setattr(e, 'input', ", ".join(input_list) if input_list else None)
+            setattr(e, 'input', ", ".join(input_list) if input_list else '')
 
         return e
 
@@ -88,9 +92,21 @@ class Workflows(resource.ResourceList):
 
     workflows = [Workflow]
 
+    def __init__(self, **kwargs):
+        self._type = 'workflows'
+
+        super(Workflows, self).__init__(**kwargs)
+
     @classmethod
     def sample(cls):
-        return cls(workflows=[Workflow.sample()])
+        workflows_sample = cls()
+        workflows_sample.workflows = [Workflow.sample()]
+        workflows_sample.next = "http://localhost:8989/v2/workflows?" \
+                                "sort_keys=id,name&" \
+                                "sort_dirs=asc,desc&limit=10&" \
+                                "marker=123e4567-e89b-12d3-a456-426655440000"
+
+        return workflows_sample
 
 
 class WorkflowsController(rest.RestController, hooks.HookController):
@@ -121,10 +137,16 @@ class WorkflowsController(rest.RestController, hooks.HookController):
             of multiple workflows. In this case they all will be updated.
         """
         definition = pecan.request.text
+        scope = pecan.request.GET.get('scope', 'private')
 
+        if scope not in SCOPE_TYPES.values:
+            raise exc.InvalidModelException(
+                "Scope must be one of the following: %s; actual: "
+                "%s" % (SCOPE_TYPES.values, scope)
+            )
         LOG.info("Update workflow(s) [definition=%s]" % definition)
 
-        db_wfs = workflows.update_workflows(definition)
+        db_wfs = workflows.update_workflows(definition, scope=scope)
         models_dicts = [db_wf.to_dict() for db_wf in db_wfs]
 
         workflow_list = [Workflow.from_dict(wf) for wf in models_dicts]
@@ -140,11 +162,18 @@ class WorkflowsController(rest.RestController, hooks.HookController):
             of multiple workflows. In this case they all will be created.
         """
         definition = pecan.request.text
+        scope = pecan.request.GET.get('scope', 'private')
         pecan.response.status = 201
+
+        if scope not in SCOPE_TYPES.values:
+            raise exc.InvalidModelException(
+                "Scope must be one of the following: %s; actual: "
+                "%s" % (SCOPE_TYPES.values, scope)
+            )
 
         LOG.info("Create workflow(s) [definition=%s]" % definition)
 
-        db_wfs = workflows.create_workflows(definition)
+        db_wfs = workflows.create_workflows(definition, scope=scope)
         models_dicts = [db_wf.to_dict() for db_wf in db_wfs]
 
         workflow_list = [Workflow.from_dict(wf) for wf in models_dicts]
@@ -157,18 +186,74 @@ class WorkflowsController(rest.RestController, hooks.HookController):
         """Delete the named workflow."""
         LOG.info("Delete workflow [name=%s]" % name)
 
-        db_api.delete_workflow_definition(name)
+        with db_api.transaction():
+            wf_db = db_api.get_workflow_definition(name)
 
-    @wsme_pecan.wsexpose(Workflows)
-    def get_all(self):
-        """Return all workflows.
+            if wf_db.is_system:
+                msg = "Attempt to delete a system workflow: %s" % name
+                raise exc.DataAccessException(msg)
+
+            db_api.delete_workflow_definition(name)
+
+    @rest_utils.wrap_pecan_controller_exception
+    @wsme_pecan.wsexpose(Workflows, types.uuid, int, types.uniquelist,
+                         types.list, types.uniquelist)
+    def get_all(self, marker=None, limit=None, sort_keys='created_at',
+                sort_dirs='asc', fields=''):
+        """Return a list of workflows.
+
+        :param marker: Optional. Pagination marker for large data sets.
+        :param limit: Optional. Maximum number of resources to return in a
+                      single result. Default value is None for backward
+                      compatability.
+        :param sort_keys: Optional. Columns to sort results by.
+                          Default: created_at.
+        :param sort_dirs: Optional. Directions to sort corresponding to
+                          sort_keys, "asc" or "desc" can be choosed.
+                          Default: asc.
+        :param fields: Optional. A specified list of fields of the resource to
+                       be returned. 'id' will be included automatically in
+                       fields if it's provided, since it will be used when
+                       constructing 'next' link.
 
         Where project_id is the same as the requester or
         project_id is different but the scope is public.
         """
-        LOG.info("Fetch workflows.")
+        LOG.info("Fetch workflows. marker=%s, limit=%s, sort_keys=%s, "
+                 "sort_dirs=%s, fields=%s", marker, limit, sort_keys,
+                 sort_dirs, fields)
 
-        workflows_list = [Workflow.from_dict(db_model.to_dict())
-                          for db_model in db_api.get_workflow_definitions()]
+        if fields and 'id' not in fields:
+            fields.insert(0, 'id')
 
-        return Workflows(workflows=workflows_list)
+        rest_utils.validate_query_params(limit, sort_keys, sort_dirs)
+        rest_utils.validate_fields(fields, Workflow.get_fields())
+
+        marker_obj = None
+
+        if marker:
+            marker_obj = db_api.get_workflow_definition_by_id(marker)
+
+        db_workflows = db_api.get_workflow_definitions(
+            limit=limit,
+            marker=marker_obj,
+            sort_keys=sort_keys,
+            sort_dirs=sort_dirs,
+            fields=fields
+        )
+
+        workflows_list = []
+
+        for data in db_workflows:
+            workflow_dict = (dict(zip(fields, data)) if fields else
+                             data.to_dict())
+            workflows_list.append(Workflow.from_dict(workflow_dict))
+
+        return Workflows.convert_with_links(
+            workflows_list,
+            limit,
+            pecan.request.host_url,
+            sort_keys=','.join(sort_keys),
+            sort_dirs=','.join(sort_dirs),
+            fields=','.join(fields) if fields else ''
+        )
